backend/requirements.txt
django==4.1.4
python-dotenv==0.21.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pillow==9.4.0
djangorestframework==3.14.0
djoser==2.1.0
xhtml2pdf==0.2.9
